Control control code to OOP
The control algorithm uses an altitude range, as well as thresholds for
velocity and acceleration in addition to methods to control altitude.
This lends itself to using a class instead of multiple global variables
and separate functions.

Updated Control.h and Control.cpp to use OOP styles
Moved old Control.h and Control.cpp to *.bak temporarily